Final Commit before deployment

+++ routes/ideas.js
@@ -1,29 +1,6 @@
 const router = require('express').Router();
 const Idea = require('../models/Idea');
 
-const ideas = [
-  {
-    id: 1,
-    text: 'Positive NewsLetter, a newsletter that only shares positive, uplifting news',
-    tag: 'Technology',
-    username: 'TonyStark',
-    date: '2022-01-02',
-  },
-  {
-    id: 2,
-    text: 'Milk cartons that turn a different color the older that your milk is getting',
-    tag: 'Inventions',
-    username: 'SteveRogers',
-    date: '2022-01-02',
-  },
-  {
-    id: 3,
-    text: 'ATM location app which lets you know where the closest ATM is and if it is in service',
-    tag: 'Software',
-    username: 'BruceBanner',
-    date: '2022-01-02',
-  },
-];
 //Get All Ideas
 router.get('/', async (req, res) => {
   try {
@@ -42,7 +19,6 @@ router.get('/', async (req, res) => {
 router.get('/:id', async (req, res) => {
   try {
     const ideas = await Idea.findById(req.params.id);
-
     res.json({
       success: true,
       data: ideas,
@@ -76,19 +52,27 @@ router.post('/', async (req, res) => {
 //Update Idea
 router.put('/:id', async (req, res) => {
   try {
-    const Updatedidea = await Idea.findByIdAndUpdate(
-      req.params.id,
-      {
-        $set: {
-          text: req.body.text || text,
-          tag: req.body.tag || tag,
+    const idea = await Idea.findById(req.params.id);
+    console.log(idea.username, req.body.username);
+    if (idea.username === req.body.username) {
+      const Updatedidea = await Idea.findByIdAndUpdate(
+        req.params.id,
+        {
+          $set: {
+            text: req.body.text || text,
+            tag: req.body.tag || tag,
+          },
         },
-      },
-      {
-        new: true,
-      }
-    );
-    res.json({ success: true, data: Updatedidea });
+        {
+          new: true,
+        }
+      );
+      return res.json({ success: true, data: Updatedidea });
+    }
+    res.status(403).json({
+      success: false,
+      msg: 'You are not Authorized to do this update',
+    });
   } catch (error) {
     res.status(500).json({ success: false, msg: error.message });
   }
@@ -97,9 +81,16 @@ router.put('/:id', async (req, res) => {
 //Delete Idea
 router.delete('/:id', async (req, res) => {
   try {
-    await Idea.findByIdAndDelete(req.params.id);
+    const idea = await Idea.findById(req.params.id);
+    if (idea.username === req.body.username) {
+      await Idea.findByIdAndDelete(req.params.id);
 
-    res.json({ success: true, data: {} });
+      return res.json({ success: true, data: {} });
+    }
+    res.status(403).json({
+      success: false,
+      msg: 'You Are not Authorized to Delete this resource',
+    });
   } catch (error) {
     res.status(500).json({ success: false, msg: error.message });
   }
